# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ProgrammeSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgrammeSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "Page[ProgrammeSummary]",
        "type": "object"
      },
...
      "ProgrammeCreate": {
        "description": "Raising a training request (FR-04).\n\nRequirements are deliberately **not** accepted here. FR-04 creates the request and\nFR-05 defines what it needs; keeping them apart is what makes the `DRAFT` →\n`REQUIREMENTS_SET` transition mean something rather than being a flag.",
        "properties": {
          "categoryId": {
            "exclusiveMinimum": 0.0,
            "title": "Categoryid",
            "type": "integer"
          },
          "endDate": {
            "format": "date",
            "title": "Enddate",
            "type": "string"
          },
          "expectedParticipants": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedparticipants"
          },
          "startDate": {
            "format": "date",
            "title": "Startdate",
            "type": "string"
          },
          "stationId": {
            "description": "Venue.",
            "exclusiveMinimum": 0.0,
            "title": "Stationid",
            "type": "integer"
          },
          "title": {
            "examples": [
              "Digital Forensics Level 2"
            ],
            "maxLength": 200,
            "minLength": 5,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "categoryId",
          "startDate",
          "endDate",
          "stationId"
        ],
        "title": "ProgrammeCreate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "ict@upf.go.ug",
      "name": "Uganda Police Force — Directorate of Human Resource Development"
    },
...
    "license": {
      "name": "Restricted — Uganda Police Force internal use"
    },
    "title": "Trainer Prediction System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/programmes": {
      "get": {
        "description": "Paginated and filterable. Search matches title **or** registry number.\n\nThe list is **not scoped by creator** — officers see all programmes, because training is coordinated across a directorate rather than owned by whoever typed the request.",
        "operationId": "list_programmes_api_v1_programmes_get",
        "parameters": [
          {
            "description": "Title or registry number.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Title or registry number.",
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          },
          {
            "in": "query",
            "name": "specializationAreaId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specializationareaid"
            }
          },
          {
            "in": "query",
            "name": "createdBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Createdby"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datefrom"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dateto"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sortby"
            }
          },
          {
            "in": "query",
            "name": "sortDir",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Sortdir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProgrammeSummary_"
                }
              }
            },
            "description": "A page of programmes."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List training programmes",
        "tags": [
          "Training programmes"
        ]
      },
      "post": {
        "description": "FR-04. Creates a programme at `DRAFT` with a generated registry number.\n\nRequirements are **not** accepted here — FR-05 defines them separately via `PUT /programmes/{id}/requirements`. Keeping the two apart is what makes the `DRAFT` → `REQUIREMENTS_SET` transition a real event.",
        "operationId": "create_programme_api_v1_programmes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgrammeSummary"
                }
              }
            },
            "description": "Programme created."
          },
          "403": {
            "description": "Only officers and administrators may raise requests."
          },
          "404": {
            "description": "Unknown category or venue."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Raise a training request",
        "tags": [
          "Training programmes"
        ]
      }
    },
    "/api/v1/programmes/{programme_id}": {
      "delete": {
        "description": "Only from `DRAFT` or `REQUIREMENTS_SET`. Anything further along has allocation history that forms part of the decision record and returns 409 — cancel it instead, which preserves the record.",
        "operationId": "delete_programme_api_v1_programmes__programme_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Deleted."
          },
          "409": {
            "description": "It has allocation history; cancel instead."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a draft programme",
        "tags": [
          "Training programmes"
        ]
      },
      "get": {
        "description": "With its lifecycle timeline, latest prediction run, and allocation count.",
        "operationId": "get_programme_api_v1_programmes__programme_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgrammeDetail"
                }
              }
            },
            "description": "The programme."
          },
          "404": {
            "description": "No such programme."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Programme detail",
        "tags": [
          "Training programmes"
        ]
      },
      "patch": {
        "description": "Blocked with 409 once the programme is `ALLOCATED` or beyond — a trainer has accepted a commitment by then, and moving the dates would silently change what they agreed to.",
        "operationId": "update_programme_api_v1_programmes__programme_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgrammeSummary"
                }
              }
            },
            "description": "Updated."
          },
          "409": {
            "description": "Too far along to edit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit a programme",
        "tags": [
          "Training programmes"
        ]
      }
    },
    "/api/v1/programmes/{programme_id}/eligibility-preview": {
      "get": {
        "description": "Runs the **gates only** and returns counts: *\"142 of 812 trainers meet these criteria.\"*\n\nCheap by construction — no scoring, no narrative. Its purpose is to let an officer discover that their criteria are too narrow **before** spending a prediction run, rather than after staring at a list of three names.",
        "operationId": "eligibility_preview_api_v1_programmes__programme_id__eligibility_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityPreviewResponse"
                }
              }
            },
            "description": "Eligible and total counts."
          },
          "409": {
            "description": "Requirements are not yet defined (FR-05)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Live eligibility preview",
        "tags": [
          "Training programmes"
        ]
      }
    },
    "/api/v1/programmes/{programme_id}/predict": {
      "post": {
        "description": "FR-06. Runs the engine over every trainer, persists the run, marks prior runs superseded, clears the re-run flag, and moves the programme to `PREDICTED`.\n\n**409 if requirements are undefined** — FR-05 forbids proceeding without them, because the engine would have nothing to match on and would rank the entire force.\n\nPrior runs are superseded rather than deleted: what the system recommended, and when, is part of the audit record.",
        "operationId": "predict_api_v1_programmes__programme_id__predict_post",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRunRead"
                }
              }
            },
            "description": "Ranking generated and stored."
          },
          "409": {
            "description": "Requirements are undefined, or no scoring policy is active."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate the ranking",
        "tags": [
          "Training programmes"
        ]
      }
    },
    "/api/v1/programmes/{programme_id}/prediction": {
      "get": {
        "description": "The latest non-superseded run: ranked candidates with their score breakdowns, rationales, confidence, and counterfactuals.\n\n**Always ordered by rank. There is no `sortBy` parameter** — BR-05 fixes the order, and letting a client re-sort would let the interface present a different recommendation from the one recorded.",
        "operationId": "current_prediction_api_v1_programmes__programme_id__prediction_get",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRunRead"
                }
              }
            },
            "description": "The current ranking."
          },
          "404": {
            "description": "No ranking has been generated yet."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The current ranking",
        "tags": [
          "Training programmes"
        ]
      }
    },
    "/api/v1/programmes/{programme_id}/requirements": {
      "put": {
        "description": "FR-05. `requiredSpecializationAreaId` is required; minimum experience and qualification are optional.\n\nMoves `DRAFT` → `REQUIREMENTS_SET`. **If a ranking already exists**, sets `requirementsChangedSincePrediction` and audits `REQUIREMENTS_CHANGED` — that flag raises the amber re-run banner, because the ranking on screen was computed against different criteria and approving from it would mean approving against requirements that no longer apply.",
        "operationId": "set_requirements_api_v1_programmes__programme_id__requirements_put",
        "parameters": [
          {
            "in": "path",
            "name": "programme_id",
            "required": true,
            "schema": {
              "title": "Programme Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgrammeSummary"
                }
              }
            },
            "description": "Requirements set."
          },
          "404": {
            "description": "Unknown discipline or qualification level."
          },
          "409": {
            "description": "The programme is past the point of redefinition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Define the requirements",
        "tags": [
          "Training programmes"
        ]
      }
    },
...
  },
...
}